% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_summary.R
\name{categorical_summary}
\alias{categorical_summary}
\title{Summarize Categorical Variables}
\usage{
categorical_summary(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A tibble with variable, category, count, and percent.
}
\description{
Provides counts and percentages for all factor or character variables.
}
\examples{
\dontrun{
categorical_summary(iris)
}
}
