% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names_simple.R
\name{clean_names_simple}
\alias{clean_names_simple}
\title{Clean and Standardize Column Names}
\usage{
clean_names_simple(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
Data with cleaned column names.
}
\description{
Clean and Standardize Column Names
}
