% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap_clean.R
\name{corr_heatmap_clean}
\alias{corr_heatmap_clean}
\title{Clean Correlation Heatmap}
\usage{
corr_heatmap_clean(data, palette = "RdBu")
}
\arguments{
\item{data}{A data frame with numeric columns.}

\item{palette}{A Brewer palette.}
}
\value{
A ggplot2 heatmap plot.
}
\description{
Clean Correlation Heatmap
}
