% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_hist.R
\name{quick_hist}
\alias{quick_hist}
\title{Quick Histogram Visualization}
\usage{
quick_hist(data, var, palette = "Blues")
}
\arguments{
\item{data}{A data frame.}

\item{var}{A numeric variable in the dataset.}

\item{palette}{A Brewer palette name.}
}
\value{
A ggplot2 histogram.
}
\description{
Quick Histogram Visualization
}
