% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cleaner.R
\name{text_cleaner}
\alias{text_cleaner}
\title{Simple Text Cleaning}
\usage{
text_cleaner(text)
}
\arguments{
\item{text}{A character vector.}
}
\value{
A cleaned character vector.
}
\description{
Simple Text Cleaning
}
